Split code into fercord_storage and fercord_bot
