+++ fercord_bot/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "fercord_bot"
+version = "0.3.0"
+license.workspace = true
+authors.workspace = true
+edition.workspace = true
+repository.workspace = true
+keywords.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+fercord_storage = { path = "../fercord_storage" }
+poise = "0.5"
+config = { version = "0.13", default-features = false, features = ["toml"] }
+serde = { version = "1", features = ["derive"]}
+serde_json = "1"
+tracing = "0.1"
+tracing-subscriber = {version = "0.3", features = ["default", "env-filter"]}
+chrono = "0.4"
+chrono-tz = "0.8"
+chrono-english = "0.1"
+reqwest = { version = "0.11", features = ["stream", "rustls-tls"]}
+clap = { version = "4", features = ["derive", "env"] }
+
+[dependencies.tokio]
+version = "1"
+features = ["rt-multi-thread", "time"]
+
+[dependencies.tokio-stream]
+version = "0.1"
+
+[dependencies.anyhow]
+version = "1"
+features = ["backtrace"]
+
+[dependencies.uuid]
+version = "1"
+features = [
+    "v4",                # Lets you generate random UUIDs
+    "fast-rng",          # Use a faster (but still sufficiently random) RNG
+    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
+    "serde",             # Add serde support
+]
