+++ fercord_bot/src/cli.rs
