+++ fercord_bot/src/discord/mod.rs
