+++ fercord_bot/src/healthchecks.rs
@@ -3,10 +3,9 @@ use chrono::Utc;
 use serde::Serialize;
 use tracing::{event, Level};
 
-use crate::healthchecks::CheckType::Database;
+use fercord_storage::prelude::*;
 
-use super::{storage::{db, kv::KVClient}};
-use super::config::DiscordConfig;
+use crate::healthchecks::CheckType::Database;
 
 #[derive(Serialize, Debug)]
 pub struct HealthCheck {
