+++ fercord_storage/Cargo.toml
@@ -1,0 +1,59 @@
+[package]
+name = "fercord_storage"
+description = "Storage crate for Fercord"
+include = [
+    "src/**/*.rs",
+    "*.md",
+    "Cargo.toml",
+    "../migrations"
+]
+license.workspace = true
+authors.workspace = true
+edition.workspace = true
+repository.workspace = true
+keywords.workspace = true
+version = "0.3.0"
+
+[features]
+default = ["sqlite"]
+sqlite = ["sqlx/sqlite"]
+postgres = ["sqlx/postgres"]
+
+[dependencies]
+poise = "0.5"
+config = { version = "0.13", default-features = false, features = ["toml"] }
+serde = { version = "1", features = ["derive"]}
+async-trait = "0.1"
+serde_json = "1"
+tracing = "0.1"
+tracing-subscriber = {version = "0.3", features = ["default", "env-filter"]}
+chrono = "0.4"
+chrono-tz = "0.8"
+chrono-english = "0.1"
+redis = { version = "0.23", features = ["tokio-comp", "json"] }
+reqwest = { version = "0.11", features = ["stream", "rustls-tls"]}
+clap = { version = "4", features = ["derive", "env"] }
+
+[dependencies.tokio]
+version = "1"
+features = ["rt-multi-thread", "time"]
+
+[dependencies.tokio-stream]
+version = "0.1"
+
+[dependencies.anyhow]
+version = "1"
+features = ["backtrace"]
+
+[dependencies.sqlx]
+version = "0.6"
+features = ["default", "chrono", "runtime-tokio-rustls", "any"]
+
+[dependencies.uuid]
+version = "1"
+features = [
+    "v4",                # Lets you generate random UUIDs
+    "fast-rng",          # Use a faster (but still sufficiently random) RNG
+    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
+    "serde",             # Add serde support
+]
