+++ fercord_storage/src/model/guild_timezone.rs
@@ -2,7 +2,7 @@ use anyhow::{anyhow, Result};
 use chrono_tz::Tz;
 use serde::{Deserialize, Serialize};
 
-use crate::storage::kv::{Identifiable, KVIdentity};
+use crate::kv::{Identifiable, KVIdentity};
 
 /// Contains the timezone set for a certain guild.
 #[derive(Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
