+++ fercord_storage/src/model/mod.rs
